fix: update to create new urlData
Before, the code written in validateToPromise only returned urlData
for a url that was already in the database. The findOrCreate fn, in
the pipeline, was biased towards the returned value of the first
argument (alt combinator checks for null/false,empty values) since
a resolved promise is always returned from findUrlAsync. Instead,
I composed a function that would take in the url from validateTo-
Promise and make that data available in the closure - for create-
UrlSync. This makes sure that if the returned value from findUrl-
Async is in fact null, then createUrlAsync with fire with the url
as an argument.